refactor: Name class text 상수화, testFixture 직관적인 이름으로 변경
