+++ src/test/java/racingcar/domain/InputDataTest.java
@@ -7,12 +7,12 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import static racingcar.domain.fixture.ConstantsFixture.CAR_NAMES;
-import static racingcar.domain.fixture.ConstantsFixture.CAR_NAME_ERROR_MESSAGE;
-import static racingcar.domain.fixture.ConstantsFixture.ROUND_COUNT;
-import static racingcar.domain.fixture.ConstantsFixture.TRY_COUNT_ERROR_MESSAGE;
-import static racingcar.domain.fixture.ConstantsFixture.WEIRD_CAR_NAME_1;
-import static racingcar.domain.fixture.ConstantsFixture.WEIRD_CAR_NAME_2;
+import static racingcar.domain.fixture.CarConstantsFixture.CAR_NAMES;
+import static racingcar.domain.fixture.CarConstantsFixture.CAR_NAME_ERROR_MESSAGE;
+import static racingcar.domain.fixture.CarConstantsFixture.ROUND_COUNT;
+import static racingcar.domain.fixture.CarConstantsFixture.TRY_COUNT_ERROR_MESSAGE;
+import static racingcar.domain.fixture.CarConstantsFixture.WEIRD_CAR_NAME_1;
+import static racingcar.domain.fixture.CarConstantsFixture.WEIRD_CAR_NAME_2;
 
 class InputDataTest {
